requirements.txt
chardet >=3.0.4
enum34;python_version<"3.4"
pyyaml >= 3.13
